+++ src/types/entry.rs
@@ -20,3 +20,11 @@ impl<T> Clone for ListEntry<T> {
 }
 
 impl<T> Copy for ListEntry<T> {}
+
+/// Resolves next link in doubly-linked list.
+#[macro_export]
+macro_rules! containing_record {
+    ($next:expr, $type:ty, $field:tt) => {
+        $next.flink.cast::<u8>().sub(memoffset::offset_of!($type, $field)).cast::<$type>()
+    };
+}