+++ CHANGELOG.md
@@ -1,2 +1,4 @@
+# Unreleased
+* Added `containing_record` macro.
 # 0.6.0
 * Began keeping changelog ig.